+++ src/loaders/GeoJsonLoader.js
@@ -1,4 +1,3 @@
-/* eslint-disable */
 import centroid from '@turf/centroid';
 import { featureCollection as turfFeatureCollection, point as turfPoint } from '@turf/helpers';
 import AbstractTwoStepLoader from './AbstractTwoStepLoader';
@@ -6,10 +5,6 @@ import { AbstractLoaderError, LoaderValidationError } from './errors/index';
 import LoaderResult from './LoaderResult';
 
 export default class GeoJsonLoader extends AbstractTwoStepLoader {
-  constructor(dataSource, params) {
-    super(dataSource, params);
-  }
-
   rejectGeoJson(reason) {
     const { url, type, fileType } = this;
     return Promise.reject(new LoaderValidationError(type, fileType, url, reason));
@@ -25,31 +20,29 @@ export default class GeoJsonLoader extends AbstractTwoStepLoader {
         if (geoJson instanceof AbstractLoaderError) {
           return Promise.reject(geoJson);
         }
-        
-        const cellsJson = {};
-        if (!(geoJson.every(cell => cell.type === 'Polygon')
-          || geoJson.every(cell => cell.type === 'Point'))) {
-          return this.rejectGeoJson('Vitessce only accepts GeoJSON that is excusively Points (i.e centroids) or Polygons');
-        }
-        geoJson.forEach((cell, index) => {
-          if (cell.type === 'Polygon') {
-            const points = turfFeatureCollection(
-              cell.coordinates[0].map(turfPoint),
-            );
-            if (cell.coordinates.length > 1) {
-              console.warn('Vitessce only accepts polygons with no holes.  Only the first ring will be used');
-            }
-            cellsJson[String(index)] = {
-              poly: cell.coordinates[0],
-              xy: centroid(points).geometry.coordinates,
-            };
-          } else {
-            cellsJson[String(index)] = {
-              xy: cell.coordinates[0],
-            };
-          }
-        });
 
+        if (!(geoJson && geoJson.type && geoJson.geometries && geoJson.type === 'GeometryCollection' && Array.isArray(geoJson.geometries))) {
+          return this.rejectGeoJson('Vitessce only accepts GeoJSON that is a GeometryCollection.');
+        }
+
+        const geoJsonGeometries = geoJson.geometries;
+        if (!geoJsonGeometries.every(cell => cell && cell.type && cell.type === 'Polygon')) {
+          return this.rejectGeoJson('Vitessce only accepts GeoJSON that is a GeometryCollection of Polygons.');
+        }
+
+        const cellsJson = {};
+        geoJsonGeometries.forEach((cell, index) => {
+          const points = turfFeatureCollection(
+            cell.coordinates[0].map(turfPoint),
+          );
+          if (cell.coordinates.length > 1) {
+            console.warn('Vitessce only accepts polygons with no holes.  Only the first ring will be used.');
+          }
+          cellsJson[String(index)] = {
+            poly: cell.coordinates[0],
+            xy: centroid(points).geometry.coordinates,
+          };
+        });
         return Promise.resolve(new LoaderResult(cellsJson, url));
       });
     return this.data;